+++ app/models/brain_test.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+# Tracks what Test My Brain tests a user has completed.
+class BrainTest < ApplicationRecord
+  # Relationships
+  belongs_to :user
+end