+++ test/models/brain_test_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class BrainTestTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end